fix: jupyterlab installed by conda (#57)
* fix: remove base containers

* fix: skip ollama to speed up build

* fix: install jupyterlab and ipykernel by conda

* fix: typo

* fix: use pip

* fix: update commit

* fix: conda install jupyter

* fix: force conda activate

* fix: skip download ollama if exists

* fix: prepare release

* add: gpu small to dev

+++ run.sh
@@ -156,6 +156,7 @@ function create_jupyter_configuration {
 }
 
 function run_jupyter() {
+	conda activate ${COOKBOOK_CONDA_ENV}
 	NB_SERVERDIR=$HOME/.jupyter
 	JUPYTER_SERVER_APP="ServerApp"
 	JUPYTER_BIN="jupyter-lab"
@@ -234,19 +235,9 @@ function conda_environment_exists() {
 function create_conda_environment() {
 	conda env create -n ${COOKBOOK_CONDA_ENV} -f $COOKBOOK_WORKSPACE_DIR/.binder/environment.yml --force
 	conda activate ${COOKBOOK_CONDA_ENV}
+	conda install jupyterlab ipykernel -y
 	pip install --no-cache-dir -r $COOKBOOK_WORKSPACE_DIR/.binder/requirements.txt
 	python -m ipykernel install --user --name "${COOKBOOK_CONDA_ENV}" --display-name "Python (${COOKBOOK_CONDA_ENV})"
-	install_base_packages
-}
-
-function update_conda_enviroment() {
-	conda activate ${COOKBOOK_CONDA_ENV}
-	conda env update -n ${COOKBOOK_CONDA_ENV} -f $COOKBOOK_REPOSITORY_DIR/.binder/environment.yml --prune
-	pip install --no-cache-dir -r $COOKBOOK_REPOSITORY_DIR/.binder/requirements.txt
-}
-
-function install_base_packages() {
-	conda install -n base -c conda-forge jupyterlab_widgets ipywidgets
 }
 
 function delete_conda_environment() {
@@ -262,7 +253,7 @@ function handle_installation() {
 		create_conda_environment
 	else
 		if { conda_environment_exists; } >/dev/null 2>&1; then
-			install_base_packages
+			echo "Conda environment already exists"
 		else
 			create_conda_environment
 		fi
@@ -270,9 +261,11 @@ function handle_installation() {
 }
 
 function start_ollama(){
-	wget "https://github.com/ollama/ollama/releases/download/v0.1.20/ollama-linux-amd64"
-	chmod 755 ollama-linux-amd64
-	mv ollama-linux-amd64 $SCRATCH/ollama
+	if [ ! -f $SCRATCH/ollama ]; then
+		wget "https://github.com/ollama/ollama/releases/download/v0.1.20/ollama-linux-amd64"
+		chmod 755 ollama-linux-amd64
+		mv ollama-linux-amd64 $SCRATCH/ollama
+	fi
 	nohup $SCRATCH/ollama serve &
 }
